Add files via upload
fixed a critical error that was introduced when used in patch 0.3.1
added additional safety features. Polymorpher will now store backups that are available in the settings menu. These Backups are persistent and will not be deleted when reverting the Character. These backups are created before anything is done that changes the actor and work independantly from the flags system used to restore actors. This provides a safe way to store the data in case another bug like the previous.