Add files via upload
"Merge Skills" will now add a bonus flat bonus if the original proficiency bonus is more then the new proficiency bonus, in the amount of the difference between the 2.
Added "Wild Shape" and "Polymorph" buttons to the dialog with predetermined settings
 - Wild Shape: keepMental, mergeSaves, mergeSkills are enabled, the rest are disabled
 - Polymorph: all settings are disabled.
These predetermined modes will nither use nor save the custom seleccted settings.
added "Keep Proficiency" to keep the class levels of the original character in case the original proficiency bonus should stay the same.
Moved the backup Character Store out of the Settings. This was overdue since the Settings were not the right place to store this information anyway. On Polymorphing the created backup will now be stored in a local file in the modules folder. The Backup Character Store interface uses this new location, but stays functionally the same.
I recommend to download any relevant data from the Character Store before updating to minimize dataloss.